+++ data-raw/data_for_regressions.R
@@ -1,12 +1,11 @@
 
-library(tidyverse)
-
 # example data from the excellent folk at https://github.com/ErikinBC/SurvSet
-example_data = read_csv("https://github.com/ErikinBC/SurvSet/raw/main/SurvSet/_datagen/output/Framingham.csv")
+example_data = readr::read_csv("https://github.com/ErikinBC/SurvSet/raw/main/SurvSet/_datagen/output/Framingham.csv")
 #example_data = rbind(example_data,example_data)
 example_data = example_data |>
-  select(-pid) |>
-  rename(sbp=num_sbp,
+  dplyr::select(-pid) |>
+  dplyr::rename(
+         sbp=num_sbp,
          dbp=num_dbp,
          scl=num_scl,
          age=num_age,
@@ -14,14 +13,20 @@ example_data = example_data |>
          sex=fac_sex,
          month=fac_month)
 
-example_data = example_data |> mutate(
-  bmi_cat = case_when(
+example_data = example_data |> dplyr::mutate(
+  
+  bmi_cat = dplyr::case_when(
     bmi <  25 ~ 0,
     bmi >= 25 & bmi < 30 ~ 1,
     bmi >= 30 ~ 2),
-  sex = as.numeric(if_else(
+  bmi_cat = haven::labelled(bmi_cat, c(Normal = 0, Overweight = 1, Obese = 2), label="BMI (categories)"),
+  
+  sex = as.numeric(dplyr::if_else(
     sex == "M", 1, 0
-  )))
+  )),,
+  sex = haven::labelled(sex, c(Female = 0, Male = 1), label="Sex")
+  
+)
 
 
 # test